Introduce custom number type
We will need a type that is capable of representing numbers in a certain
way. Right now we foresee two important qualities this type needs to
have:
1) It needs to be suitable for working with currencies. For one, that
   means there should not be any loss in precision as is the case when
   dealing with floating point values. Second, our rounding algorithm,
   which we will need eventually, needs to be fair.
2) We need custom parsing logic to correctly create number instances
   from a string, from JSON, and potentially from integers and floats.

This change marks a first step in this direction by introducing a new
type, Num, which is based off of num_rational's BigRational. Right now
we don't support much more than parsing from a string and converting
back into one, with the conversion back to string still using the
default one of BigRational.