Add support for fairly rounding Num values
When dealing with currencies we ultimately need to round to a sensible
number of decimal places, commonly two. It is ultimately advisable to do
so in a fair manner, i.e., without any bias (as would be the case if we
always rounded up or down, for example).
This change adds support for rounding Num values in such a manner. In
particular, we follow the "banker's rounding" algorithm, by which we
round towards the closest integer as usual and to the closest even one
in case of the fractional part being 0.5.