Fix parsing of small numbers in Num
As it turns out our Num struct fails to parse values between 0 and 1
properly. The reason, once more, can be found in the weird sign logic
that BigInt employs. Because we used the numer's sign and it is NoSign,
and because NoSign causes all data passed in to be discarded, we ended
up with an invalid denom.
This change fixes the problem by special casing for NoSign.