Avoid using fixed size intermediate integers
As part of the conversion from a string into a Num we parse the
pre-decimal point position into a fixed size integer and the
post-decimal point position one as well. However, it is very well
possible that either actually exceeds the range of the fixed size
integer.
With this change we do it right and use a BigInt for this intermediate
value as well. That also has the added benefit that we can no longer
overflow the power operation.