Add support for deserialization from ints and floats
We initially started off working under the assumption that a Num will be
created from a string, given that the first crate we used this crate in
represented serialized numbers as strings. However, that will not always
be the case.
To be more flexible and future proof, this change also adds support for
creating a Num by deserializing an integer or floating point value.