Do not automatically round in Display implementation
We initially implementing the logic to convert a Num into a string
representation we opted for automatically rounding the value to two
decimal places. That was done 1) because it was "good enough" and 2)
because it was simpler in terms of implementation.
With this change we revisit this decision and no longer preform any
rounding of values when converting them into a string, as rounding of
this sort should be explicit in user code. We use the long division
algorithm for converting a rational into its decimal representation.