Cap the maximum precision to use when formatting
So far we have ignored one potential specific issue which is the fact
that some rational numbers have to be represented as an infinite
character sequence when displayed as a decimal number. 1/3 is such an
example. The problem we have in this case is that our rounding code
would run forever (or rather cause a stack over flow on actual
hardware).
To fix this problem, we cap the maximum precision to use, such that
formatting will terminate. Note that we still allow the user to
overwrite this maximum by means of the precision formatting syntax
("{:.xx}").
Lastly, by virtue of introducing such an upper limit we now also round
unconditionally when converting a number into a string, so as to ensure
that we don't just cut off decimal positions that we ignore.