Make serde dependency optional
We strive to keep our list of dependencies as small as possible. One
dependency that is not strictly required by all clients and usual made
optional is that to serde.
With this change we introduce a feature for it (disabled by default)
that pulls in serde and provides proper serialization and
deserialization support.