Add support for mixed integer arithmetic
Depending on the use case it can come quite handy to be able to perform
arithmetics not only between Num objects but also with raw integers.
This change adds support for such operations. Note that we currently
only support arithmetics with anything that can be converted into an
i32. There is no technical reason why we can't inter operate with, say,
u64, support for that just has not been implemented yet.