Honor format precision more exactly
Upon initial introduction of the functionality we have treated the
format precision as an upper bound on the number of decimals to print.
However, that is not how the standard library formats floating point
values: it guarantees to print as many decimals as the user asked for,
and not potentially less.
With this change we adjust our behavior to do the same: instead of
stopping when the remaining decimals are zeros we just keep going if the
user asked for a specific precision.