Add methods for conversion into integers
In an attempt to make it easier to work with Num objects in all sorts of
contexts, this change introduces additional methods that allows for
conversion into different integer types. The newly added method
to_integer can be used to convert into a BigInt as provided by the
num-bigint crate. to_i64 and to_u64 on the other hand, allow for
conversion into a built in integer type, if the value fits.