Provide a custom Debug implementation for Num
So far we have used a derive to provide an implementation of Debug for
the Num struct. However, the result is an extremely verbose output,
because the call is effectively just forwarded to BigRational, which
ends up printing the nominator and denominator, both of which being
BigInt objects, which in turn have a very verbose representation.
We are not interested in debugging BigRational itself and so this change
provides a custom implementation of Debug that results in a much more
concise output comprised of a fraction.