Provide From implementations for all integer types
As it turns out BigInt does have constructors for creating objects from
the various integer types. That means that we do not have to use the
poor man's solution we currently have which requires an i32 in order to
safely use the from_slice constructors.
As a first step in the transition away from this kludge, this change
introduces From implementation for all known integer types.