Remove new_bigint helper function
When we assumed that there was no way to convert arbitrary integers into
a BigInt, we created a helper function, new_bigint, that accomplishes
the job for anything convertible into an i32.
Given that we can create a BigInt from an arbitrary integer, this
function is no longer useful. Hence, with this change we remove it.