Correctly report the underlying error source for ParseNumError
The std::error::Error trait which should generally be implemented for
most if not all error types has a source method that can be used to peek
into an underlying error. By default, this method does not report any
such error.
We, however, may have such an error when the parsing of an integer
failed. Hence, with this change we correctly implement the source method
for ParseNumError to take care of this reporting.