Introduce support for specifying a minimum precision to use
It appears that Rust's default formatting options are insufficient for
some of the use cases our Num type will be used in. In particular, in
some applications it may be required that we display at least a certain
number of post decimal positions, but optionally more if precise
specification requires so. E.g., financial applications may have a
requirement to display two such positions, but individual values may
require more if the desired value cannot be represented precisely in
only two.
This change adds support for such use cases by introducing a "display
adapter" that is essentially a builder-style object that can be
formatted as usual and which allows for controlling the minimum
precision.