Support unstable channel in CI Debian image
Debian, which we use as the distribution of choice in our CI pipeline,
has recently removed kcov from the testing channel. As such, our
pipeline now fails in the code coverage stage, because it can't install
kcov.
This change fixes this problem by adding the unstable channel. It also
removes the 'deb-src' entry, which is not necessary because we do not
install anything from source.