Rename NumVisitor to HumanReadable
As it turns out our deserializer does not work with binary encoding such
as bincode. The reason for that is that by relying in the
deserialize_any method for driving the visitor we implicitly require a
self describing format. Most binary formats do not fit into this
category.
In the future we intend to add another deserializer that will be used
when deserializing binary data and that does not use deserialize_any. In
preparation for this introduction, this change renames the existing
NumVisitor to HumanReadable.