Add support for serializing/deserializing using bincode
This change adds support for serializing to and deserializing from
non self-describing formats such as bincode for the Num type. We use a
relatively easy to encode (and decode) binary representation of our
internally used BigInt objects. While the result is still not optimal in
terms of binary size, encoding overhead is decreased significantly over
JSON (and our to string conversion; times are for a full round trip
comprised of serialization and deserialization of the result):

  test bench_bincode ... bench:         277 ns/iter (+/- 0)
  test bench_json    ... bench:       7,411 ns/iter (+/- 22)