Add support for working with version 0.3 num-* crates
The num-bigint and num-rational crates have released new versions
containing breaking changes. Because we expose types from at least
num-bigint in the public API surface, we have to consider bumping this
dependency in a backwards incompatible way as a breaking change
ourselves. At the same time, a breaking change in this crate would mean
a large updating and releasing effort for dependent crates.
To circumvent this problem, this change introduces a new feature,
num-v03, that can be used to request the 0.3 versions of the num-*
crates. Furthermore, we enable another feature, num-v02, by default,
which preserves the existing behavior.