Rename string conversion tests
We have a bunch of tests that check that we can create Num objects from
various numbers encoded as strings. Their names, however, seem to imply
that we are converting from an integer directly.
With this change we append the string '_str' to the name of each test in
order to make the intended purpose clearer.