Add TryFrom conversions between number types
This change introduces conversion methods from Num into "smaller" ones.
In general, these conversions cannot be assumed to be lossless. As such,
the are fallible. For example, there is no guarantee that we can convert
a Num into a Num64, because the former has unlimited precision while the
latter does not.