Tag more functionality with #[inline]
Except for generic code, Rust generally does not inline functionality
across crate boundaries. This crate's API surface is somewhat large, but
at the same time it mostly just wraps existing methods on inner types --
making it arguably important that we allow for inlining to happen.

To that end, this change tags more of crate's function with #[inline],
allowing the compiler to inline them at usage sites.