Introduce From conversions from Num{32,64} to Num
In the same way that it may become necessary to convert from Num to
Num32/Num64, it may also be a requirement to go the other way around,
from Num32 or Num64 to Num.
While that could always be done manually, it is unnecessarily cumbersome
to do. This change makes that possible more easily by introducing the
corresponding From implementation.