Switch to using cargo llvm-cov for code coverage assessment
Our code coverage collection runs have repeatedly failed because of a
what appears to be a bug in the actions-rs/tarpaulin implementation.
This action, it turns out, has actually been (soft?) deprecated [0].
Attempting to switch to installing cargo-tarpauling directly takes ages,
and we already know that we would rather avoid handling caching in
GitHub Actions, because of their API mess.
However, it appears as if cargo llvm-cov may be usable by now, and is
generally said to provide more accurate coverage information. Thus, with
this change we switch over to using it for collection of code coverage.
Note that cargo-tarpaulin also seems to support using LLVM coverage as a
backend, but it still suffers from the aforementioned issues.

[0] https://github.com/actions-rs/tarpaulin/issues/6