Generate docs.rs documentation with num-v04 and serde features
By default docs.rs uses the default set of a crate's features when
generating the documentation. In our case, we would like to have it
generated with the num-v04 and serde features enabled, which are
non-default.
This change adds the necessary configuration to the Cargo.toml file to
make that happen.